+++ packages/signers/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @mysten/signers
 
+## 0.2.7
+
+### Patch Changes
+
+- 8356318: Fix broken message signing for the Ledger signer
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [7d66a32]
+- Updated dependencies [eb91fba]
+- Updated dependencies [19a8045]
+  - @mysten/sui@1.29.0
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [9a94aea]
+  - @mysten/sui@1.28.2
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [3cd4e53]
+  - @mysten/sui@1.28.1
+
 ## 0.2.3
 
 ### Patch Changes